okex-python-sdk-api/requirements.txt
attrs==19.3.0
bleach==3.1.5
certifi==2020.4.5.2
chardet==3.0.4
docutils==0.16
idna==2.9
keyring==21.2.1
more-itertools==8.4.0
packaging==20.4
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.8.2
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.3
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
tqdm==4.46.1
twine==3.1.1
urllib3==1.25.9
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==0.57.0
